Refactor de tipo de dato en Id de entidades; Avances sobre el issue #4

+++ GTRetail-ejb/src/main/java/ar/com/gtsoftware/model/BaseEntity.java
@@ -36,23 +36,23 @@ public abstract class BaseEntity implements Serializable {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Basic(optional = false)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Version
     private Integer version;
 
-    public BaseEntity(Integer id) {
+    public BaseEntity(Long id) {
         this.id = id;
     }
 
     public BaseEntity() {
     }
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
